+++ src/main/java/com/javachina/service/CommentService.java
@@ -9,16 +9,16 @@ import com.javachina.model.Comment;
 
 public interface CommentService {
 	
-	public Comment getComment(Integer cid);
+	public Comment getComment(Long cid);
 		
-	public Comment getTopicLastComment(Integer tid);
+	public Comment getTopicLastComment(Long tid);
 	
 	public List<Comment> getCommentList(QueryParam queryParam);
 	
 	public Page<Comment> getPageList(QueryParam queryParam);
 	
-	public boolean save( Integer uid, Integer toUid, Integer tid, String content, Integer createTime );
+	public boolean save( Long uid, Long toUid, Long tid, String content, Long createTime );
 	
-	public boolean delete(Integer cid);
+	public boolean delete(Long cid);
 		
 }
